requirements.txt
langchain 
langchain-community 
langchainhub 
langchain-chroma  
langchain-openai 
langgraph 
feedparser 
openai 
requests
pytz
scikit-learn
gcsfs
google-cloud-firestore
pypdf
flask
flask-cors
Flask-Caching
IPython
python-dotenv